+++ jest.config.build.js
@@ -1,4 +1,4 @@
-module.exports = require('@socifi/jest-config')(undefined, {
+module.exports = require('@socifi/jest-config')('build', {
     transformIgnorePatterns: [
         '<rootDir>/dist',
     ],