+++ rollup.config.js
@@ -1,9 +1,9 @@
-import getConfig from 'socifi-rollup-config';
+import getConfig from '@socifi/rollup-config';
 import path from 'path';
 import babel from 'rollup-plugin-babel';
 import replace from 'rollup-plugin-re';
 import packageJson from './package.json';
-const { getBaseBabelConfig } = require('socifi-rollup-config/src/helpers');
+const { getBaseBabelConfig } = require('@socifi/rollup-config/src/helpers');
 
 const configs = getConfig(packageJson, path.resolve(__dirname, 'src'));
 
@@ -22,6 +22,14 @@ export default configs.map((config) => {
                         {
                             test: 'rest-api-handler/src',
                             replace: 'rest-api-handler/dist',
+                        },
+                        {
+                            test: 'cookie-api-handler/src',
+                            replace: 'cookie-api-handler/dist',
+                        },
+                        {
+                            test: 'gpx-builder/src',
+                            replace: 'gpx-builder/dist',
                         },
                     ],
                 });